Remove Debug.* logging mecahnism and just use Debug.WriteLine
The Debug.tracef mechanism was slow because it was using sprintf even in
Release DLLs.  This was showing up big in perf traces of project loading

Replacing with System.Diagnositcs.Debug.WriteLine. Keep things simple.
Also because we're reusing the core services functionality more and
removing the dependency on a diagnostics technique is now sensible

Specific kinds of messages can be detected with grep etc....