Remove global setting for language version choice
The lower part of the services API was relying on a global MonoDevelop
configuration parameter. Instead use F# 3.0 if it is available (that has
been the default anyway). This will make it easier to move this lower
part into a separate DLL which can be used from different editors
(planned by @tpetricek). In future we can pick up the language version
from a project file setting, so we pass the setting through as a
parameter.

Also add a 4-second delay in starting processing in the F# background
compiler worker in LanguageService.fs. This improves project-load
performance substantially.