Implement Goto-definition (meta-D on mac, F12 on Windows)
This uses the F# compiler capability for GetDeclarationLocation to
implement goto-definition.

It is fiddly. MonoDevelop's view of the world is in terms of symbols
(e.g. for renaming, refactory etc.) The F# compiler doesn't report
symbols, it just reports data for operations like goto definition and
get-data-tip-tet.

We resolve language items to an under-specified LocalResolveResult
(with an additional field for the data tip text). Goto-definition can
operate on the LocalResolveResult.

The commands are Meta-D on mac, F12 on windows. We may have to add
these for F# iles - they work as keystrokes, but do not appear in any
menu when editing an F# file.