[MSBuild] support optional project type migrations
Allow migrations to be optional, and have
migration handlers return a CIL Type instead
of bool. This allows for migrations to not only
migrate their flavor GUIDs but also to register
new CIL project types.

Finally allow migrations to handle the prompting
for allowing a migration to actually happen. If
no prompt override is provided by the migration
handler, the default will be shown (ignore,
migrate, backup and migrate).

Optional migrations that are ignored by the user
will simply load with the old flavor GUID whereas
required migrations (default behavior) will fail
to load when ignored by the user.