Fix size issue in pad toolbar buttons
The problem was that a size request for the pad toolbar was made
before adding the pad to the dock container. This is a problem because
the size depends on the style, and the style may change when a widget
is added to a container. The fix is to avoid a call to CalcNewSizes
which is making the SizeRequest call.
This fixes bug 8102 - Buttons on autohidden toolbox pad have wrong
size until selected