Improvements in the welcome page
The welcome page is not a document view anymore. It is now shown as an
overlay on top of the dock frame.

Changed the way branding is supported. Instead of loading the welcome
page contents from an xml file, the welcome page is now a widget that
can be replaced by an add-in. The content of the welcome page is added
by the subclass of WelcomePageWidget. This allows more flexibility in
the customization of the widget.