[Ide] Disable threaded rendering for now as it's causing crashes
Pango is not yet threadsafe on any platform so it is not safe to access
pango types from a background thread. It corrupts memory in static hashtables
causing either crashes or memory corruption on windows. The same could be
happening on Mac and Linux.

The threaded rendering for the Dock widget causes this, and potentially more
places.