[Ide] Add an overload with an optional topic parameter to ShowDocs.
When showing the docs from the addons infrastructure, we want to be
able to not only add the docs to the tree, but also open directly to
them. We can do that using MacDoc/WinDoc's topic URLs, so allow
passing one here (this is the entry point that that infrastructure uses).