[VersionControl] Don't clear the old VersionInfo object too early
We should only get rid of the old object when the new one is ready.
This prevents issues where a modified file would temporarily be
marked as unmodified while the asynchronous update is happening.
With this change we retain the old VersionInfo object until the
new one is ready, thus removing the need to create a temporary
VersionInfo object when calling OnGetVersionInfo.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=8718