[RazorTemplating] Fix null reference exceptions in code completion
Razor Generator made the assumption that the host would not be
re-used so it did not re-create the parser on each parse. This broke
the design-time background parse, causing unhandled exceptions to be
thrown on certian incomplete constructs.

Instead of a Parser property, we how have a ParserFactory property, and
the parser has been re-written so it doesn't pass state via a dictionary
on the host, but instead plugs into the Razor code generation
mechanisms to alter the generated code directly, as other directives do.