[Project] Ensure SolutionItem.BoundToSolution fired correctly
It's possible during deserialization for SolutionItems to be added to a
ParentFolder that is not yet connected to a ParentSolution. In those
cases we should avoid firing BoundToSolution, and instead recursively
dispatch BoundToSolution when the ParentSolution does become connected.

Fixes an exception viewing the Components folder on projects in
Solution Folders.