[Build] Deuglify buildinfo in About dialog
Make sure the all lines have keys, instead of trying to add one
to the first line later. This fixes an issue where the Mac build
had been altered to prepend a line with a key, causing the About
dialog to display a line with two keys and a line with no key.