[Razor] Re-sign Razor assemblies to they load on Windows
We were bundling a copy of the Razor assemblies that had been
built on Mono, so they used the MS key token but were not
actually validly signed, so .NET refused to load them.

Since these are used for code completion only, and not for user
projects, we can change the key via ildam/ilasm.