Remove duplicates in GetAll methods.
When we build in Wrench, for some unexplained reason, the System
keychain is included twice. This stops us from building the MacInstaller
because we fail to find the correct signing certificate.

The fix is to use HashSet instead of List when iterating through all the
items so that duplicates are not included.