[Core] Don't re-scan known frameworks from every runtime
This saves us from reading a lot of assembly names. On my machine,
with 7 Mono runtimes (2.10.9, 3.0.0, 3.0.1, 3.0.3, MonoAndroid,
MonoTouch, Mono from source) this cuts the number of assembly name
reads from 1493 to 118.

with a warm disk cache, this saves ~50ms on MD startup time. With a
cold cache it would be a lot more.