[Core] More fine-grained lock for assembly name cache
This means in general we'll be able to return cached names even
while some other thread is reading an assembly's name from disk.
In the worse case we may read a single assembly's name from disk
multiple sives if several threads request the name for that
assembly at the same time and it is not yet cached.