Fix excetion on tool tip
Previously we returned just the raw summary is an empty string was
returned.  The problem is an empty string is returned from
GetDocumentationMarkup if the phrase "To be added" is found by the
GetDocumentationMarkup method.  See:
https://github.com/mono/monodevelop/blob/master/main/src/core/MonoDevelo
p.Ide/MonoDevelop.Ide.TypeSystem/AmbienceService.cs

The best thing to do here is simple return the emty sting as we have no
documentation anyway.