ProjectFileSelectorDialog: allow adding new files
Internally store filters as a list of
SelectFileDialogFilter instances instead of
retaining only the patterns.

Add support for adding new files directly to the
project from the dialog, conservatively
restricting the filters and build actions on the
open file dialog to match those of the
ProjectFileSelectorDialog.

This should help eliminate confusion for
adding icon files to applications via the plist
editor. Previously .icns files would have to
already be added to the project in order to be
selected for use.