Use the new concurrent sgen mark-sweep when running with make run
It's a stupid hack, but at the moment there's no easy way to
concatenate that env var as mono does not allow trailing separators
when it parses the option string.

Older versions of mono also bail out when they encounter options
they don't understand, so this is the simplest way of probing for
'marksweep-conc' support and falling back to regular sgen if it's
not found.