[Core] Register core frameworks in ascending order
Workaround for the broken SystemAssemblyService.GetTargetFrameworkForAssembly
which marks package assemblies as targeting the first framework with
the referenced mscorlib. It was marking 4.0-referencing assemblies as
targeting 4.5, which broke build for 4.0 projects including MD itself.
By registering 4.0 first we fix this build problem, though it may now think
that assemblies are compatible with 4.0 when they actually require 4.5.