[MSBuild] Put all our builders in an isolated directory
We cannot place our builder executables in the same directory as
other binaries. If we do this, the normal assembly loading rules
can result in assemblies like Mono.Cecil.dll being loaded from
the private monodevelop build dir instead of from the 'correct'
place.

For example if there is a custom msbuild dll which defines some
tasks used by the build process, and that dll references Mono.Cecil.dll
and supplies it's own copy of that library, we need to load *that* one
and not the one in monodevelops private bin dir. The assembly loading
rules check the application base dir before checking in the same dir
as the custom dll.