[MSBuild] Put all our builders in an isolated directory
We cannot run our builder executables from a directory which contains
other assemblies as the standard assembly resolution rules mean that
we would end up loading MonoDevelops private copies of some binaries
instead of the copies bundled with libraries implementing msbuild tasks.

For example if there is a custom msbuild dll which defines some
tasks used by the build process, and that dll references Mono.Cecil.dll
and supplies it's own copy of that library, we need to load *that* one
and not the one in monodevelops private bin dir. The assembly loading
rules check the application base dir before checking in the same dir
as the custom dll.

Note that we now create a proper v4.0 builder as MonoDevelop requires
that there is a MonoDevelop.Projects.Formats.MSBuild.exe file in
it's main directory, but we cannot actually use this one for building.