[Debugger] Implement our own "TreeRowReference" since gtk#'s is buggy(?)
For some reason, row.Valid() is returning true, but somehow row.Path
is null. This seems to happen when rows are deleted above the path
that the row references, making TreeRowReference basically useless.

Not sure if it's a bug in our binding or gtk+ itself, or if I'm just
somehow using it wrong(?), but my custom implementation seems to work.