Add optional expansion of environment variable macros within hint paths
MSBuild supports a wide variety of macros within all aspects of the build
file.  MonoDevelop's internal build system generally does not support
these macros.  This commit adds support for a particular special case,
where the macro simply references an environment variable.  This feature
could cause problems in projects that use a macro name defined within the
project, that just happens to match an environment variable in a user's
environment, because with MSBuild the macro would use the project variable,
whereas this implementation would use the environment variable.  To deal
with this possibility, this behavior is optional, defaults to off, and can
be enabled within the Build options panel.