Added multi-file+folder commit support.
This allows you to select multiple navigation nodes and commit them. This is an extension of PR #286,
and enhances support for it.

But here comes a question: Should I remove the CommandUpdateHandler?
This command can get too CPU intensive in some cases, because of recursive directory searches.
We have the MessageBox on operations that cannot be done. I'll add another commit if we need this.