[NUnit] Add support for NUnitLite style Xml output
NUnitLite changed the format of it's xml, but the new formats are both
a little buggy. The best format is supposed to be like nunit2. The main
difference is that it no longer writes <test-suite type="Namespace" elements
in the xml. The other difference is that it does not write the namespace
qualified type names. This means by default you cannot tell which namespace
a test comes from.

I fixed this by modifying MacUnit to always write the full type name as well
as modifying the MonoDevelop parser to cope with the lack of 'Namespace'
elements. This seemed like the simplest approach.