WIP: Implement an ultra basic alternative to the existing debug tooltip logic
This test demonstrates the mkrueger approved method of getting the information
we need for the debugger tooltips.

The code in 'ResolveExpression' is what the debug tooltips should use to figure
out the expression that needs to be passed to the debugger. It seems to work
for the few cases I've added to the testsuite but it's by no means comprehensive
test suite yet! :p