[NUnit] Initialize the Test Results pad immediately on test start
If we asynchronously initialize this then it's possible we will
abort the test run because the previous run was cancelled and
the 'cancelled' variable may not have been re-initialized back
to 'false'.

Fixes the bug where if a test run was cancelled the next time you
try to run the tests would fail.