[Debugger] Fixed code-completion for the ExpressionEvaluatorDialog
Fixes bug #12804

Also made it so that you can open the dialog w/o any text selected. It was
driving me nuts that I kept clicking the "Expression Evaluator" menu item
and the dialog wasn't showing.

IMHO, if we don't want to show it with an empty selection, then it should
be disabled when an expression isn't selected in the text editor.