[NUnit] Add better support for streaming updates from mdtool/guiunit
If we get streaming results from these test runners we need to clear
out the test results pad before we parse the results xml. Some information
is only available in the results xml (currently) so to avoid having each
test appear twice in the output pad we clear it out after we receive
all the streaming updates and then re-populate it from the results
xml file.