[MacPlatform] Remove some duplicated NSApplication.Init calls
We need to enforce that NSApplication is only initalised at application
startup. We still call NSApplication.Init more than once and we still
do it lazily which is Bad. We can end up (re)initializing on the wrong
thread and break monomac.

This kills the worst of the issues and should allow MonoDevelop to run
again.