[MacPlatform] Rely on Xwt.Mac initializing MonoMac for us
We don't want to call NSApplication.Init multiple times as MonoMac is now
going to explicitly prevent this by throwing an exception. It's a bug if
we ever do this, especially if we try to delay-init monomac and end up
calling NSApplication.Init either on the wrong thread or after we've
already tried to use monomac types.