[TestRunner] Do not reference guiunit at all
The test runner cannot reference guiunit by binary or project reference
as if  we use any types from the 'guiunit' assembly we will not be able
to run any tests from a test assembly which references a different
guiunit.

The runner was written specifically to dynamically load guiunit.exe
at runtime and access the required types via reflection to avoid this.