[Git] Visual advice to comply with conventions
This is a tiny hook in the VersionControl.Git project that allows the
CommitDialog show a visual advice to the user when they surpass
the character count recommended for the first line of a Git commit
message.

The text that surpasses the limit is marked as red, and a tooltip
for the textview is shown. This is inspired by a visual indicator
that Vim already has.