[Ide] Fix DockBarItem size requests
Set the requisition to the primary surface size rather than adding it,
otherwise we can get a widget that grows in size each time it receives
a size request. This can occur when we change themes, or on Windows when
any number of system settings (such as desktop background) are changed.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=15967