Fix over-eager symbol parsing
Previously if a tooltip was requested at the start of an ident after
a dot such as the 'c' in:

List.collect

Then the '.' would be detected as a symbolic operator fragment.
Because we start parsing in the middle of idents/symbolic operators
it seems difficult to do better than this. The only tooltips that
won't display now are if they are requested just after a dot in
a symbolic operator.