[MSBuild] Default file format can no longer be changed
There is almost no reason to make it easy for users to default
to creating older file formats, all it does is increase chance
of messing things up. Projects can still be exported, loaded and
roundtripped with older formats, so we're not removing
functionality.

We now always create projects with the MSBuild12 file format,
which can be opened in VS2010 SP1 and later.