[tests] Disable the testing of Gtk sharp projects in our test suite
We cannot reliably run that test as our test harness ends up hitting
many different shutdown race conditions and so the fixture frequently
get's marked as a failure.

Disable the initialization of the xamarin studio services and also the
testing of Gtk sharp projects until someone can spend the time to fix
all the races.