Overhauled FSharpInteractibe to take into account various edge conditions in Windows.
Removed debug diagnostics as commands are already logged in MonoDevelop.
Inherited from abstract base class to save implementing useless update
command.
Altered InteractiveCommand so that’s FSharpInteractive is not available
then the commands are visible but disabled.  This can happen because of
various edge conditions if the addin is reinstalled etc.