[NuGet] Custom draw the entire list view row in Add Packages dialog.
Custom draw allows the background color of the entire row to be set.

When no items are checked then the current item has a blue strong
selection color. When one or more items are checked the selection
color is changed to grey.

Centered NuGet package icon.
Using images for check boxes.
Check boxes are semi-transparent until one or more is checked.
Then the check boxes are made opaque.