[NuGet] Add proxy support.
Use MonoDevelop credential provider if username/password not
in NuGet.config file.

Prompt for proxy credentials when accessing NuGet repository in
Add Packages dialog.

Prompt for credentials for NuGet repositories requiring
authentication which is not stored in NuGet.config file.

Do not prompt for credentials in preferences just show the error next
to the package source.