[NuGet] Fix package source timing out on Windows.
Repeatedly switching between the package sources in Preferences and
the Add Packages dialog would stop returning any packages.

The problem seems to be that .NET has a limit on the number of
concurrent http requests to the same url and the code was not
disposing the request causing this limit to be hit. Once hit
all requests would never be sent and eventually time out.