[VersionControl] Disable the 'QueryThreadWorks' test.
This test is racy and cannot give reliable results. It is possible
that the query associated with any 'AddDirectory' or 'AddFile' call
can actually be processed before we call the next GetVersionInfo
or GetDirectoryInfo call.

The test relies on being able to add a file/directory but not
retrieve it during the next call to GetDirectoryVersionInfo. It
assumes that a Thread.Sleep will be needed to wait for the query
to be processed, but that is not the case.