[NuGet] Show package in Packages folder when installing.
Previously when a package was installing, after being selected in the
Add Packages dialog, only the status area would have a visual
indication that the package was installing. It was considered that
this was perhaps not obvious. A large NuGet package can take a while
to download and it may be confusing for a user to not see any package
in the Solution pad.

Now when the install is initiated the package is shown in the Packages
folder in the Solution pad. The package id is highlighted in red to
show it is not yet installed. Once installed the package id is shown
in normal black text.

Issues:

When a package fails to install the package is removed from the
Packages folder. Still need to investigate how this could be improved
such as giving the user a chance to reinstall the package again
from the Solution pad.

Need a different icon to show the package is not installed. Perhaps
a different icon (animated) for the current state of being installed
which is different to the unrestored package state.