[NuGet] Improve Package Console logging.
Make multiple package actions easier to read by separating them by
a new line. Using new line instead of Visual Studio's use of '=' and
'-' characters since a new line looks cleaner when multiple actions
log only a little information individually.

Move PowerShell and package license checks inside the package action
so that information about this can be logged alongside the rest of the
package information. Previously these checks were done at the start
and all the information was logged together for all packages. This
code will need to be revisited if Xamarin Studio supports PowerShell
and license acceptance in the future.

Log a starting message for each package action so the Package Console
is not empty whilst a large package is being downloaded. Need to
investigate if the download can be logged and possibly integrated into
the status area's progress similar to how Visual Studio works.
into logging this download information and possibly