[NuGet] Prevent package updates to unpublished/unlisted package.
Currently the latest Newtonsoft.Json is version 6.0.1 but on the
official NuGet feed there is another unpublished version 6.0.2.
The Add Packages dialog would not show 6.0.2 but the user could use
the Update Packages command to update to 6.0.2 even though it has not
been published.

Fix is to prevent finding packages which are unpublished/unlisted.