Simplify background parsing/typechecking
This avoids a lot of needless reparses. All these reparses make the navigation extension nervous. Prior to this change, the nav bar would typically change three times when switching document.
It seems Parse() is designed to be long running, and it won't lock the UI. So there's no need to do this trickery with requesting reparses.