[NuGet] Fix invalid cast exception.
Addin was trying to cast Project to DotNetProject which may fail.

One way to reproduce this is to have a SharedProject in the solution
along with a normal DotNetProject and try to add a NuGet package from
the Add Packages dialog to the DotNetProject.

Since the NuGet addin only supports DotNetProjects then all other
project types are ignored.