Fix #350: Avoid caching ToolTipText failures:
When the internal timeout is hit (when the timeout works) a ToolTipText of '(loading description...)' is shown. Using lazy here means that even when retrying later, this will still be shown.
I made it a function that looks up the result every time. I looked at the corresponding code in FSC, and it will cache the result.