[NuGet] Reload files open in text editor if NuGet changes them.
Fixes:

#18911 - packages.config file does not reload itself if it is open in the XS text editor and need to be closed and reopened again
https://bugzilla.xamarin.com/show_bug.cgi?id=18911

#18932 - web.config/ app.config not get automatically reloaded when changed by nuget
https://bugzilla.xamarin.com/show_bug.cgi?id=18932

Any files that are changed by NuGet are saved and then when NuGet
has finished processing the NuGet addin calls
FileService.NotifyFilesChanged, passing all the files that were
changed.

FileService.NotifyFilesChanged handles reloading or warning the user
that the file has changed and needs to be reloaded.

Should possibly look at the NuGet addin being able to update the file
directory in the text editor instead of updating it directly on disk.