[NuGet] Remove 'Check Package Console' message from Package Console.
Fixes: #18926 - 'Check Package Console for details' error is
displayed both in XS Status bar and PackageConsole

https://bugzilla.xamarin.com/show_bug.cgi?id=18926

When an error or warning occurs when running a NuGet action, such as
installing a NuGet package, then the status bar has a general message
indicating that there was a problem and more information can be found
in the Package Console window. This general error/warning message is
also duplicated in the Package Console window when it is not helpful
to the user in this location and it is also redundant. This message is
of the form:

"Could not add NUnit. Please see Package Console for details"

The general error/warning message shown in the status bar is now no
longer displayed in the Package Console.

Warning messages that are specific to NuGet are still logged at the
end of the Package Console as a summary so the user does not have to
search through the entire Package Console output, which can be
length, to find the warnings themselves. However the general status
bar warning message is not repeated in the Package Console.

Note that success messages that are shown in the status bar when a
NuGet package action completes (e.g. 'NUnit succesfully added') are
still shown in the Package Console. This duplication of the message
has been left since it summarises that the action has completed when
the user is only looking at the Package Console. These success
messages also never refer to the Package Console.