[NuGet] Fix package id hyperlinks for recently installed packages.
Fixes 19134
https://bugzilla.xamarin.com/show_bug.cgi?id=19134

If the package is returned from an online NuGet feed (e.g. nuget.org)
that has a gallery url link then the package id should have a
hyperlink. If the package is from a local directory on your machine
then it will not have a hyperlink. For packages in local directory
Xamarin Studio does not necessarily know they are from NuGet.org or
from an online feed since there is no information in the package
itself about where it came from. Visual Studio has the same behaviour
for packages in a local directory.

If the Add Packages dialog is first opened then all the package id
hyperlinks are OK. If you install a package then the next time the
Add Packages dialog is opened a hyperlink for the recently used
package was not being displayed when it should be.