Fix 18429 - Package description pane overscroll appearance
https://bugzilla.xamarin.com/show_bug.cgi?id=18429

On overscrolling the package information on the right hand
side of the Add Packages dialog the grey background of
the dialog was shown and the left hand line border would
separate from the bottom of the dialog. Now the border
does not separate and the grey background of the dialog
is not shown.