[NuGet] Show user changes made to packages.config in Solution pad.
Fixes Bug 19143 - Deleted packages from package.config file should
remove from Project/Package without reopening solution.

https://bugzilla.xamarin.com/show_bug.cgi?id=19143

Use the FileService.FileChanged event to trigger the Solution pad
to refresh the Packages node based on what is in the packages.config
file. Using the FileChanged event allows the packages.config file
to be edited in the text editor and have changes in Solution pad
appear when the file is saved. It also allows the packages.config
file to be edited outside of Xamarin Studio and have the changes
reflected in the Solution pad when switching back to Xamarin
Studio.