[Ide] Fix crash after disabling addin that extends the Solution pad.
Fixes 19149:
https://bugzilla.xamarin.com/show_bug.cgi?id=19149

Disabling an addin that extends the solution pad with custom nodes
that have custom menus and then trying to use those menus will crash
Xamarin Studio with an error similar to:

ERROR [2014-04-18 16:53:32Z]: An unhandled exception has occured. Terminating MonoDevelop? True
System.InvalidOperationException: Extension node not found in path: /MonoDevelop/PackageManagement/ContextMenu/ProjectPad/PackageReference
  at Mono.Addins.ExtensionContext.GetExtensionObjects (System.String path, System.Type arrayElementType, Boolean reuseCachedInstance) [0x0001f] in source/monodevelop/main/external/mono-addins/Mono.Addins/Mono.Addins/ExtensionContext.cs:670
  at Mono.Addins.ExtensionContext.GetExtensionObjects (System.String path, Boolean reuseCachedInstance) [0x00000] in source/monodevelop/main/external/mono-addins/Mono.Addins/Mono.Addins/ExtensionContext.cs:558
  at MonoDevelop.Components.Commands.CommandManager.CreateCommandEntrySet (Mono.Addins.ExtensionContext ctx, System.String addinPath) [0x00007] in source/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Components.Commands/CommandManager.cs:222
  at MonoDevelop.Ide.Gui.Components.ExtensibleTreeView.CreateContextMenu () [0x000e2] in source/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui.Components/ExtensibleTreeView.cs:1800
  at MonoDevelop.Ide.Gui.Components.ExtensibleTreeView.ShowPopup (Gdk.EventButton evt) [0x00002] in source/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui.Components/ExtensibleTreeView.cs:1776
  at MonoDevelop.Components.ContextMenuTreeView.OnButtonPressEvent (Gdk.EventButton evnt) [0x00052] in source/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Components/ContextMenuTreeView.cs:62
  at Gtk.Widget.buttonpressevent_cb (IntPtr widget, IntPtr evnt) [0x00016] in gtk-sharp-2.12.21/gtk/generated/Widget.cs:2938

To reproduce this, open a project and add NuGet packages to it.
Expand the packages node so the package references are visible in the
Solution pad. Disable the NuGet addin using the Addin Manager. Then
try to right on a package reference and use one of the context menus.

This problem also occurs if you open a GTK# project, disable the
GTK# Visual Designer, then try to use one of the context menus
for the User Interface node in the Solution pad.

When an addin is unloaded any extension paths it uses are sent to
anybody who is interested by the AddinManager's ExtensionChanged
event. The SolutionPadCodon was not matching the project pad
extensions created by the NuGet addin on the path
'/MonoDevelop/Ide/Pads/ProjectPad'.

Match attempted:

    ExtensionChangedEventArgs.Path: "/MonoDevelop/Ide/Pads/ProjectPad"
    codonpath: "/MonoDevelop/Ide/Pads/ProjectPad/ProjectPad"

SolutionCodon:

    Path: /MonoDevelop/Ide/Pads/ProjectPad
    Id: ProjectPad

The fix was to stop appending the Id to the path when trying to match
since the path already includes this.