[NuGet] Improve error handling when password encryption fails.
Fix 19257 - Exception on updating NuGet source
https://bugzilla.xamarin.com/show_bug.cgi?id=19257

Check that mono can encrypt package source passwords before
applying the changes. Encryption can fail if the
~/.config/.mono/keypairs directory has the wrong permissions.
If the encryption fails then the user is prompted instead of being
shown an exception dialog.

The keypairs directory can be created by mono with the wrong
permissions by the HttpWebRequest when a response is received from
a url that uses https. There is another bug logged for this:

https://bugzilla.xamarin.com/show_bug.cgi?id=19274

To fix the 'Improperly protected user's key pairs in '~/.config/
.mono/keypairs'' error message the following command can be run to
fix the directory permissions:

chmod u=rwx,go= ~/.config/.mono/keypairs