Only show 'Implement Interface' when cursor is on an interface.
Seems to work ok. You type 'interface IDisposable with' and then hover over IDisposable.
You don't have to type the 'interface ' or the ' with' either. We'll need to check for incomplete forms of this when generating code.
Ideally, you could just type 'IDisposable' and apply the refactoring to that.

The command still does nothing though.