[NuGet] Remove package search from search bar if project not selected.
Fixes 19277 - Unable to search package using Xamarin Studio search bar
https://bugzilla.xamarin.com/show_bug.cgi?id=19277

The package search 'Search Packages...' can be used from the unified
search window to open the Add Packages dialog and search for a NuGet
package. Previously the package search was available in unified search
pop up window if a solution was not open and nothing would happen if
it was selected. NuGet packages can currently only be added to a
project so instead of not showing the Add Packages dialog the package
search is removed from the unified search pop up window when a
solution is not open or a project is not selected.