[Ide] Fix repeating folder name issue
When the last file of a folder is removed, MD adds an empty folder object
to the project file list, so that the folder is shown in the solution pad.
However, that code was not checking if the folder did exist. It does it now.
This fixes bug #19285 - Shared Asset projects are repeating folders.