[NuGet] Fix package source preferences incorrectly showing warning.
Fixes 19308 - 'Incorrect Credentials' warning is not always up to date.
https://bugzilla.xamarin.com/show_bug.cgi?id=19308

In Preferences->Packages->Sources, entering the correct username and
password for an existing package source that is showing 'Invalid
credential' warning does not remove the warning. This warning would
disappear on closing and re-opening the package source preferences.

The credentials being used to verify the source can be accessed were
the last credentials saved in the NuGet.config and not the ones
currently held in memory whilst the package source preferences were
being changed.

Once a valid username and password has been entered then NuGet will
use this credential even though the user may change the username and
password to incorrect values. The user would need to restart Xamarin
Studio for the incorrect values to take affect. This is because NuGet
has its own caching mechanism.

The package source preferences is not affected by NuGet caching since
any credentials applied by NuGet to the first request are removed to
force NuGet to use the latest credentials.